Modify setup and add requirements
