+++ requirements.txt
@@ -1,0 +1,20 @@
+numpy>=1.20.3
+matplotlib 
+scipy>=1.7.0 
+seaborn 
+math 
+emcee 
+corner 
+scikit-learn 
+mpl_toolkits 
+cycler 
+statistics 
+time 
+warnings 
+bilby 
+typing 
+pathlib
+pathlib
+tempfile
+subprocess
+logging
